[WIP] Adding a new autofixable no-unused-import-rule
To add this rule, we has to upgrade eslint, and therefore to fix errors from the new rules added by the new version.

We also had to migrate to the new eslint configuration system and this is where everything blew up.
IN PROGRESS.

This commit does this for the play container.